Actually track sessions connected to pipelines.
The level of abstraction in the API is wrong.
We should have sessions, problems, and solutions.
One session contains zero or more problems, which
contains zero or more solutions.  As it is there's
an extra layer of indirection which isn't really expressed
and so can't be manipulated by the TA3.